Add latencymon for disco alarmas in global report
