+++ src/main.js
@@ -21,8 +21,7 @@ Vue.use(LibraryDelayer, {
   libraries: {
     ripe_widget_api: "https://stat.ripe.net/widgets/widget_api.js",
     latencymon_widget: [
-      "https://www-static.ripe.net/static/rnd-ui/atlas/static/measurements/widgets/latencymon/dev/libs/require.min.js",
-      "https://www-static.ripe.net/static/rnd-ui/atlas/static/measurements/widgets/latencymon/latencymon-widget-main.js"
+      "https://atlas.ripe.net/resource/latencymon/latencymon-widget-main.js"
     ],
     tracemon_widget: [
       "https://www-static.ripe.net/static/rnd-ui/atlas/static/measurements/widgets/tracemon/dev/libs/require.min.js", //preload require
