+++ src/views/charts/tables/DiscoAlarmsTable.vue
@@ -24,17 +24,18 @@
             </a>
           </div>
           <div v-else>
-              {{getCountryName(props.row.streamname)}}
+              {{countryName(props.row.streamname)}}
           </div>
         </q-td>
-        <q-td key="starttime"> {{props.row.starttime}} </div> </q-td>
-        <q-td key="endtime"> {{props.row.endtime}} </div> </q-td>
+        <q-td key="starttime"> {{dateFormatter(props.row.starttime)}} </q-td>
+        <q-td key="duration"> {{duration(props.row.starttime, props.row.endtime)}} </q-td>
         <q-td key="deviation">{{props.row.avglevel}}</q-td>
+        <q-td key="nbdiscoprobes"> {{props.row.nbdiscoprobes}} </q-td>
       </q-tr>
       <q-tr v-show="props.expand" :props="props">
           <q-td colspan="100%" class="IHR_nohover" bordered>
             <div v-if='props.expand' class="IHR_side_borders">
-                Some very fancy view
+            <latencymon :start-time="halfDayShift(startTime, -1)" :stop-time="halfDayShift(stopTime, 1)" :msm-prb-ids="msmPrbIds(props.row.discoprobes)" style="max-width: 93%; margin: 0 auto;"/>
             </div>
           </q-td>
         </q-tr>
@@ -45,259 +46,13 @@
 
 <script>
 import CommonTableMixin from './CommonTableMixin.vue';
-
-const isoCountries = {
-    'AF' : 'Afghanistan',
-    'AX' : 'Aland Islands',
-    'AL' : 'Albania',
-    'DZ' : 'Algeria',
-    'AS' : 'American Samoa',
-    'AD' : 'Andorra',
-    'AO' : 'Angola',
-    'AI' : 'Anguilla',
-    'AQ' : 'Antarctica',
-    'AG' : 'Antigua And Barbuda',
-    'AR' : 'Argentina',
-    'AM' : 'Armenia',
-    'AW' : 'Aruba',
-    'AU' : 'Australia',
-    'AT' : 'Austria',
-    'AZ' : 'Azerbaijan',
-    'BS' : 'Bahamas',
-    'BH' : 'Bahrain',
-    'BD' : 'Bangladesh',
-    'BB' : 'Barbados',
-    'BY' : 'Belarus',
-    'BE' : 'Belgium',
-    'BZ' : 'Belize',
-    'BJ' : 'Benin',
-    'BM' : 'Bermuda',
-    'BT' : 'Bhutan',
-    'BO' : 'Bolivia',
-    'BA' : 'Bosnia And Herzegovina',
-    'BW' : 'Botswana',
-    'BV' : 'Bouvet Island',
-    'BR' : 'Brazil',
-    'IO' : 'British Indian Ocean Territory',
-    'BN' : 'Brunei Darussalam',
-    'BG' : 'Bulgaria',
-    'BF' : 'Burkina Faso',
-    'BI' : 'Burundi',
-    'KH' : 'Cambodia',
-    'CM' : 'Cameroon',
-    'CA' : 'Canada',
-    'CV' : 'Cape Verde',
-    'KY' : 'Cayman Islands',
-    'CF' : 'Central African Republic',
-    'TD' : 'Chad',
-    'CL' : 'Chile',
-    'CN' : 'China',
-    'CX' : 'Christmas Island',
-    'CC' : 'Cocos (Keeling) Islands',
-    'CO' : 'Colombia',
-    'KM' : 'Comoros',
-    'CG' : 'Congo',
-    'CD' : 'Congo, Democratic Republic',
-    'CK' : 'Cook Islands',
-    'CR' : 'Costa Rica',
-    'CI' : 'Cote D\'Ivoire',
-    'HR' : 'Croatia',
-    'CU' : 'Cuba',
-    'CY' : 'Cyprus',
-    'CZ' : 'Czech Republic',
-    'DK' : 'Denmark',
-    'DJ' : 'Djibouti',
-    'DM' : 'Dominica',
-    'DO' : 'Dominican Republic',
-    'EC' : 'Ecuador',
-    'EG' : 'Egypt',
-    'SV' : 'El Salvador',
-    'GQ' : 'Equatorial Guinea',
-    'ER' : 'Eritrea',
-    'EE' : 'Estonia',
-    'ET' : 'Ethiopia',
-    'FK' : 'Falkland Islands (Malvinas)',
-    'FO' : 'Faroe Islands',
-    'FJ' : 'Fiji',
-    'FI' : 'Finland',
-    'FR' : 'France',
-    'GF' : 'French Guiana',
-    'PF' : 'French Polynesia',
-    'TF' : 'French Southern Territories',
-    'GA' : 'Gabon',
-    'GM' : 'Gambia',
-    'GE' : 'Georgia',
-    'DE' : 'Germany',
-    'GH' : 'Ghana',
-    'GI' : 'Gibraltar',
-    'GR' : 'Greece',
-    'GL' : 'Greenland',
-    'GD' : 'Grenada',
-    'GP' : 'Guadeloupe',
-    'GU' : 'Guam',
-    'GT' : 'Guatemala',
-    'GG' : 'Guernsey',
-    'GN' : 'Guinea',
-    'GW' : 'Guinea-Bissau',
-    'GY' : 'Guyana',
-    'HT' : 'Haiti',
-    'HM' : 'Heard Island & Mcdonald Islands',
-    'VA' : 'Holy See (Vatican City State)',
-    'HN' : 'Honduras',
-    'HK' : 'Hong Kong',
-    'HU' : 'Hungary',
-    'IS' : 'Iceland',
-    'IN' : 'India',
-    'ID' : 'Indonesia',
-    'IR' : 'Iran, Islamic Republic Of',
-    'IQ' : 'Iraq',
-    'IE' : 'Ireland',
-    'IM' : 'Isle Of Man',
-    'IL' : 'Israel',
-    'IT' : 'Italy',
-    'JM' : 'Jamaica',
-    'JP' : 'Japan',
-    'JE' : 'Jersey',
-    'JO' : 'Jordan',
-    'KZ' : 'Kazakhstan',
-    'KE' : 'Kenya',
-    'KI' : 'Kiribati',
-    'KR' : 'Korea',
-    'KW' : 'Kuwait',
-    'KG' : 'Kyrgyzstan',
-    'LA' : 'Lao People\'s Democratic Republic',
-    'LV' : 'Latvia',
-    'LB' : 'Lebanon',
-    'LS' : 'Lesotho',
-    'LR' : 'Liberia',
-    'LY' : 'Libyan Arab Jamahiriya',
-    'LI' : 'Liechtenstein',
-    'LT' : 'Lithuania',
-    'LU' : 'Luxembourg',
-    'MO' : 'Macao',
-    'MK' : 'Macedonia',
-    'MG' : 'Madagascar',
-    'MW' : 'Malawi',
-    'MY' : 'Malaysia',
-    'MV' : 'Maldives',
-    'ML' : 'Mali',
-    'MT' : 'Malta',
-    'MH' : 'Marshall Islands',
-    'MQ' : 'Martinique',
-    'MR' : 'Mauritania',
-    'MU' : 'Mauritius',
-    'YT' : 'Mayotte',
-    'MX' : 'Mexico',
-    'FM' : 'Micronesia, Federated States Of',
-    'MD' : 'Moldova',
-    'MC' : 'Monaco',
-    'MN' : 'Mongolia',
-    'ME' : 'Montenegro',
-    'MS' : 'Montserrat',
-    'MA' : 'Morocco',
-    'MZ' : 'Mozambique',
-    'MM' : 'Myanmar',
-    'NA' : 'Namibia',
-    'NR' : 'Nauru',
-    'NP' : 'Nepal',
-    'NL' : 'Netherlands',
-    'AN' : 'Netherlands Antilles',
-    'NC' : 'New Caledonia',
-    'NZ' : 'New Zealand',
-    'NI' : 'Nicaragua',
-    'NE' : 'Niger',
-    'NG' : 'Nigeria',
-    'NU' : 'Niue',
-    'NF' : 'Norfolk Island',
-    'MP' : 'Northern Mariana Islands',
-    'NO' : 'Norway',
-    'OM' : 'Oman',
-    'PK' : 'Pakistan',
-    'PW' : 'Palau',
-    'PS' : 'Palestinian Territory, Occupied',
-    'PA' : 'Panama',
-    'PG' : 'Papua New Guinea',
-    'PY' : 'Paraguay',
-    'PE' : 'Peru',
-    'PH' : 'Philippines',
-    'PN' : 'Pitcairn',
-    'PL' : 'Poland',
-    'PT' : 'Portugal',
-    'PR' : 'Puerto Rico',
-    'QA' : 'Qatar',
-    'RE' : 'Reunion',
-    'RO' : 'Romania',
-    'RU' : 'Russian Federation',
-    'RW' : 'Rwanda',
-    'BL' : 'Saint Barthelemy',
-    'SH' : 'Saint Helena',
-    'KN' : 'Saint Kitts And Nevis',
-    'LC' : 'Saint Lucia',
-    'MF' : 'Saint Martin',
-    'PM' : 'Saint Pierre And Miquelon',
-    'VC' : 'Saint Vincent And Grenadines',
-    'WS' : 'Samoa',
-    'SM' : 'San Marino',
-    'ST' : 'Sao Tome And Principe',
-    'SA' : 'Saudi Arabia',
-    'SN' : 'Senegal',
-    'RS' : 'Serbia',
-    'SC' : 'Seychelles',
-    'SL' : 'Sierra Leone',
-    'SG' : 'Singapore',
-    'SK' : 'Slovakia',
-    'SI' : 'Slovenia',
-    'SB' : 'Solomon Islands',
-    'SO' : 'Somalia',
-    'ZA' : 'South Africa',
-    'GS' : 'South Georgia And Sandwich Isl.',
-    'ES' : 'Spain',
-    'LK' : 'Sri Lanka',
-    'SD' : 'Sudan',
-    'SR' : 'Suriname',
-    'SJ' : 'Svalbard And Jan Mayen',
-    'SZ' : 'Swaziland',
-    'SE' : 'Sweden',
-    'CH' : 'Switzerland',
-    'SY' : 'Syrian Arab Republic',
-    'TW' : 'Taiwan',
-    'TJ' : 'Tajikistan',
-    'TZ' : 'Tanzania',
-    'TH' : 'Thailand',
-    'TL' : 'Timor-Leste',
-    'TG' : 'Togo',
-    'TK' : 'Tokelau',
-    'TO' : 'Tonga',
-    'TT' : 'Trinidad And Tobago',
-    'TN' : 'Tunisia',
-    'TR' : 'Turkey',
-    'TM' : 'Turkmenistan',
-    'TC' : 'Turks And Caicos Islands',
-    'TV' : 'Tuvalu',
-    'UG' : 'Uganda',
-    'UA' : 'Ukraine',
-    'AE' : 'United Arab Emirates',
-    'GB' : 'United Kingdom',
-    'US' : 'United States',
-    'UM' : 'United States Outlying Islands',
-    'UY' : 'Uruguay',
-    'UZ' : 'Uzbekistan',
-    'VU' : 'Vanuatu',
-    'VE' : 'Venezuela',
-    'VN' : 'Viet Nam',
-    'VG' : 'Virgin Islands, British',
-    'VI' : 'Virgin Islands, U.S.',
-    'WF' : 'Wallis And Futuna',
-    'EH' : 'Western Sahara',
-    'YE' : 'Yemen',
-    'ZM' : 'Zambia',
-    'ZW' : 'Zimbabwe'
-};
-
+import Latencymon from "@/components/ripe/Latencymon";
+import getCountryName from "@/plugins/countryName.js";
 
 export default {
   mixins: [CommonTableMixin],
   components: {
+    Latencymon
   },
   props: {
     data: {
@@ -353,11 +108,11 @@ export default {
           sortable: false
         },
         {
-          name: "endtime",
+          name: "duration",
           required: true,
-          label: "Reconnection Time",
+          label: "Duration (minutes)",
           align: "left",
-          field: row => row.endtime,
+          field: row => this.duration(row.starttime, row.endtime),
           format: val => val,
           sortable: true
         },
@@ -370,17 +125,35 @@ export default {
           format: val => val,
           sortable: true
         },
+        {
+          name: "discoProbes",
+          required: true,
+          label: "Nb. Disco. Probes",
+          align: "left",
+          field: row => row.nbdiscoprobes,
+          format: val => val,
+          sortable: true,
+        },
       ],
     };
   },
   methods: {
-      getCountryName (countryCode) {
-        if (isoCountries.hasOwnProperty(countryCode)) {
-            return isoCountries[countryCode];
-        } else {
-            return countryCode;
-        }
+      duration(start, end){ 
+        let durationMin = Math.ceil(Math.abs(new Date(end) - new Date(start)) / (1000*60));
+        return durationMin
       },
+      dateFormatter(datetime){ 
+        var dt = new Date(datetime)
+        var options = { year: 'numeric', month: 'long', day: '2-digit', hour: '2-digit', minute: '2-digit',  timeZone: 'UTC' };
+        return dt.toLocaleDateString(undefined, options)
+      },
+      countryName(code){
+        return getCountryName(code)
+      },
+      msmPrbIds(probes){ 
+        var probeIds = probes.map( probe => { return probe.probe_id });
+        return {1030:probeIds, 1031:probeIds}
+      }
   },
 };
 </script>
